GRF-371: fix CategoryTree instead of Category
